+++ R/saturate.R
@@ -1,6 +1,6 @@
 #' Make a color more saturated
 #'
-#' @details The colors will be trainsformed to HSL color space (hue, saturation,
+#' @details The colors will be transformed to HSL color space (hue, saturation,
 #'  lightness) where the saturation of the color will  be modified.
 #'  The saturation of a color takes a value between 0 and 1, with 0 being black
 #'  and 1 being white. The \code{shift} argument takes a value between 0 and 1,
@@ -35,13 +35,13 @@ clr_saturate <- function(col, shift = 0.5) {
   hsl <- decode_colour(col, to = "hsl")
   hsl[, 2] <- pro_transform(hsl[, 2], 100, shift)
 
-  rgb <- farver::convert_colour(hsl, "hsl", "rgb")
-  color(encode_colour(rgb_norn(rgb)))
+  rgb <- convert_colour(hsl, "hsl", "rgb")
+  color(encode_colour(rgb_norm(rgb)))
 }
 
 #' Make a color more desaturated
 #'
-#' @details The colors will be trainsformed to HSL color space (hue, saturation,
+#' @details The colors will be transformed to HSL color space (hue, saturation,
 #'  lightness) where the saturation of the color will  be modified. The
 #'  saturation of a color takes a value between 0 and 1, with 0 being black and
 #'  1 being white. The \code{shift} argument takes a value between 0 and 1,
@@ -76,6 +76,6 @@ clr_desaturate <- function(col, shift = 0.5) {
   hsl <- decode_colour(col, to = "hsl")
   hsl[, 2] <- pro_transform(hsl[, 2], 0, shift)
 
-  rgb <- farver::convert_colour(hsl, "hsl", "rgb")
-  color(encode_colour(rgb_norn(rgb)))
+  rgb <- convert_colour(hsl, "hsl", "rgb")
+  color(encode_colour(rgb_norm(rgb)))
 }